Re-organize jsx component structure
previously, the jsx components for each webpage were farther away
from where they were used that makes it quite difficult to navigate
and hurts readability. This new directory structure keeps all of
the jsx components in the same directory as their respective webpage
components. The webpage components should be named 'index.jsx'.
