+++ src/pages/home/Officers.jsx
@@ -3,7 +3,7 @@ import React from 'react';
 const Officer = ({name, position, desc, img, linkedIn, github}) => (
 	<div className="officer">
 		<div className="officer__avi-container">
-			<img src={require(`../images/officers/${img}.jpg`)} alt={`${name}'s Avatar`} />
+			<img src={require(`../../images/officers/${img}.jpg`)} alt={`${name}'s Avatar`} />
 		</div>
 		<h3>{name}</h3>
 		<h4>{position}</h4>
@@ -17,7 +17,7 @@ const Officer = ({name, position, desc, img, linkedIn, github}) => (
 					href={`https://www.linkedin.com/in/${linkedIn}/`}
 					title={`${name} on LinkedIn`}
 				>
-					<img src={require('../images/logos/media/linkedin-social.svg')} alt="LinkedIn logo" />
+					<img src={require('../../images/logos/media/linkedin-social.svg')} alt="LinkedIn logo" />
 				</a>
 				{github && (
 					<a
@@ -27,7 +27,7 @@ const Officer = ({name, position, desc, img, linkedIn, github}) => (
 						href={`https://github.com/${github}`}
 						title={`${github} on GitHub`}
 					>
-						<img src={require('../images/logos/media/github-social.svg')} alt="GitHub Logo" />
+						<img src={require('../../images/logos/media/github-social.svg')} alt="GitHub Logo" />
 					</a>
 				)}
 			</div>
